Allow running xfstests in parallel
Create the xfstests configuration file with sections and allow to run a
specific section on the vm.

When multiple sections are specified the required number of VMs are
clones off of the VM that was going to facilitate the testing and each
section is ran in it's own VM in parallel.

Signed-off-by: Lukas Czerner <lczerner@redhat.com>